Rewrite snare in Rust.
This does everything the Python version can and a bit more (notably it can send
emails when a per-repo program fails containing its stderr/stdout). However, the
intention is to become quite a bit more sophisticated in the future, so the
architecture of this is more complicated than it really needs to be at the
moment.

Architecturally this system is split in two:
  * The `httpserver` listens for incoming hooks, checks that they're valid, and
    adds them to a `Queue`.
  * The `jobrunner` pops elements from the `Queue` and runs them in parallel.

These two components run as two different threads: the `httpserver` writes a
solitary byte to an "event pipe" to wake up the `jobrunner` when the queue has
new elements. This allows the `jobrunner` to use a single interface for listen
for completed jobs as well as new jobs.

Since this is intended to be a fairly traditional Unix daemon (albeit written in
Rust), most of the system goes out of its way to keep running even in the face
of annoying errors. There are still a few things that can cause it to panic, but
those can be reduced/removed in the future.