Remove the pointlessly paranoid "always check the queue every minute".
This paranoia was entirely pointless, because if it was ever necessary, we would
simply get stuck elsewhere. The idea was that, perhaps, sometimes we might not
be notified that there's a job on the queue so we should (infrequently) check
the queue even if we haven't been told there's an update. However, this could
only happen if poll() failed. If that happened, then our ability to read
stderr/stdout from child processes would also fail, at which point we'd be
stuffed. But poll() isn't very likely to fail, so the paranoia is completely
pointless: we might as well simplify the code and be done with it!

While here, simplify the condition still further, and prevent snare from waking
up unnecessarily (when there are jobs on the queue, but no spare slots to run
them).