Allow per-repository settings.
This change allows users to specify most settings on a per-repository basis. The
README contains more details, but the most important part can be seen in this
example configuration file:

  port = <port>
  github {
    reposdir = "<path>"
    match ".*" {
      email = "abc@def.com"
    }
    match "a/b" {
      email = "ghi@jkl.com"
    }
  }

match blocks contain a regex matching against a repository name, and are
evaluated in order from top to bottom with each successful match overriding
previous settings. This allows users to specify defaults which are only
overridden for specific repositories. In this example the repository `c/d` will
have errors sent to `abc@def.com` but `a/b` will have errors sent to
`ghi@jkl.com`.

This was a more involved change than I had hoped, in part because it changed the
authentication mechanism. In short, in order to authenticate a webhook request,
we have to load the data, find the relevant config (since `secret` can now be
specified per-repository), and then authenticate.