RepoConfigs no longer use ownership sharing, but are moved.
Although sharing allowed us to slightly minimise memory usage, the saving was
illusory: we still had to `clone()` `String`s and so on later. One possibility
is to cache `RepoConfig`s (and distribute them through `Arc` or similar), but
that seemed unnecessarily fussy, and would also mean we'd have to implement
extra stuff like cache eviction.

This commit thus simplifies things: every time we query the `Config` about a
repo, we get back a new `RepoConfig` that is not bound in anyway to `Config`.
However, since the GitHub secret is a `SecStr`, my assumption is that a) it's
expensive to clone (since it's doing `mprotect()` and so on) b) duplicating it
repeatedly throughout the heap might make it easier for an attacker to find and
decode it. We thus return that seaprately from the `RepoConfig`.