Handle the user changing the number of maxjobs after SIGHUP.
If the user asks for more jobs to be run, we have an easy task: if they ask for
fewer to be run, then it is much trickier. The approach this commit takes for
the latter case is simple, but means that we can find ourselves in situations
where we are not ever able to actually reduce the number of maximum jobs that
are running.