Add queue kinds.
For any given repository R, snare previously ran all requests in parallel if
possible. This commit allows jobs to be run in parallel or sequentially (with
the latter being the default). "Sequentially" means "only run one job per
repository at a time": other jobs will be left on the queue and popped in FIFO
order when one has finished.