Add a 'user' global option.
This is almost entirely useful for starting snare as root: we now refuse to run
as root (unless the user specifies `user="root"`, at which point they're
explicitly telling snare that they're happy with the risk) and require a `user`
to change into. The `README` documents the precise actions we undertake as part
of the change of user.