Log via syslog.
We log via LOG_DAEMON (on OpenBSD this means that messages show up in
/var/log/daemon and on Debian in /var/log/daemon.log) unless "-d" was specified
in which case we output to stderr.

There isn't a nice Rust API for these functions, so we have to use the libc
crate and put up with `unsafe` blocks. I was surprised to realise that OpenBSD's
man page for openlog doesn't make clear what happens to `ident` parameter.
Debian's man page is clearer on this matter:

  The  argument  ident in the call of openlog() is probably stored as-is.  Thus,
  if the string  it  points  to  is  changed,  syslog()  may  start prepending
  the changed string, and if the string it points to ceases to exist, the
  results are undefined.  Most portable is  to  use  a  string constant.

We thus deliberately leak memory in order to do "the right thing" (and on Debian
and OpenBSD doing "the wrong thing" causes random data to be spewed to syslog,
which certainly backs up the man page statement above!).