Fail authentication if a secret is present.
Previously we authenticated successfully if a signature was present, but if the
GitHub side shared a secret, and our side didn't, we let it through anyway. In
one sense this is fine, but it means that one can easily forget to specify a
secret (and when I say "one", I mean "I have already made this mistake").

This commit thus returns UNAUTHORIZED if: the request is signed and we don't
have a secret; we have a secret and the request isn't signed; or our secret
doesn't match the signed request.