We no longer need to do a string check on canonicalised paths.
github_valid_username and github_valid_reponame provide precisely the guarantees
we need to make sure that the owner and repo names are safe to use as part of
pathnames. The string check on canonicalised paths was always a hack (and,
if you encountered symlinks, or similar, confusing) approximation to these
better checks.