For the event/owner/repo validations, explicitly log violations.
Since there are no absolute guarantees that these validations will always match
GitHub's rules, we log an error if input isn't validated, so that the subsequent
bug reports will make some sense to us. My guess is that this is most likely for
the 'event' validation, but only time will tell.