Report the repository name when sending email errors.
At the moment, if a job fails users get an email with a subject "snare error" --
it can then be a bit difficult to work out what repository the error in the
email body refers to. This commit changes it so that the email subject "snare
error when running <repo>".