Move to tokio 1.
The last time I tried this, it failed, as some dependencies still relied on
tokio < 1, and we ended up with two tokio runtimes being created, with the
obvious outcome that nothing worked. It now seems that all the dependencies are
now working with tokio >= 1, so we can finally move forwards. I've also had to
do "cargo update", so there's lots of obfuscating churn in `Cargo.lock`: the
real changes are in `Cargo.toml`.