Remove hyper & tokio; implement an HTTP server; and add system tests.
This commit is a thorough yak shave. First I tried to upgrade hyper,
which we previously used for snare's HTTP server. However, it has
thoroughly changed its API and no longer provides an HTTP server
directly (there is an external crate which is supposed to work, but I
couldn't get it to compile). That slowly made me rethink snare's HTTP
server, because the use of hyper is why snare also required tokio,
with all that comes with it (async, lots of dependencies, and the
overloading of my pea brain).

So, second, I removed hyper and tokio and implemented my own
minimalistic HTTP server (heavily modelled on pizauth). This hugely
reduces our dependencies (from 213 -> 172 for `cargo build`) and reduces
the final binary size by 20% in release mode.

However, how do we know that taking on the burden of running our own
HTTP server has left snare correct? One of snare's long-term weaknesses
is that we haven't have good system-level tests, and I've often been
lacking confidence when working on snare that my changes haven't broken
something important.

So, third, and finally, this commit adds system level tests is
`tests/basic.rs`. The most thorough of these fires a full GitHub-esque
request at snare, forcing it to authenticate, and run a test command.
While it's not perfect, it's a lot better than nothing.

This commit has also survived being run against GitHub itself. There is
more to do (e.g. the tests currently bind to fixed ports; and there
aren't enough system tests), but this feels like a sufficiently big unit
of work in and of itself.