Don't bind to a fixed port in tests.
Previously we used fixed ports for tests. If someone else happened to be
using those ports, then the test failed. This commit has test listening
on port 0, which causes the OS to allocate an unused port. In order that
we can connect, we need snare to tell us what port it's bound to, a
feature which is gated behind an `_internal_testing` feature.

It's quite hard to integrate the old `bad_config` test into this new
framework (my attempt ended up with some surprising type inference
errors). Since it's testing something that's well captured elsewhere
(the unittests in `Config`), I've removed it rather than make the test
runner even more complex than it now is.