Move common test behaviour to its own module.
As we add more tests, dumping everything in one file is going to become
increasingly hard to understand. However, each file in the `tests/`
directory is its own crate, which makes factoring out helper code
difficult. Fortunately, there's a special hack in Cargo to ignore
`tests/*/mod.rs` so we can put common behaviour in such a file and not
have it treated as a test. This commit does just that.