Move away from the `log` crate.
In retrospect, moving to this was a mistake: it adds lots of dependencies
(including one that has a security advisory), but we can easily replicate
its features locally, and do so in a way that doesn't rely on implicit
global variables.

This commit therefore is a sort-of manual reversion of 45666ba0 (an actual
revert is impossible as too much has changed since then), but also very
slightly improves the situation that pertained before 45666ba0.