Handle both json and urlencoded requests.
I singularly failed to realise that GitHub requests can come with a
Content-Type of JSON or urlencoded: previously snare only handled the
latter. Handling both automatically is fairly easily, but doing so
highlighted that the tests incorrectly said they had JSON content-type
but were actually urlencoded. Since JSON content-type is easier to
understand, the tests now consistently use that, except for a (new) test
to ensure that we really handle urlencoded correctly.