Simplify timeout checks and remove a race condition.
This commit simplifies how we calculate when `poll` should return.
Whilst doing this I realised that the HTTP server check is in the wrong
place and is subject to a race condition: we could empty the queue, and
then get stuck not processing anything for an arbitrary period of time.